Add separate modes for testing and downloading
A few options have been added to the command line allowing for a test mode, where the input JSON file is read, parsed and its content printed back. In download mode, the JSON input file is parsed and all the associated files are download from the METASPACE AWS cloud repository.